# Conflicts:
#	iotdb-client/cli/src/main/java/org/apache/iotdb/cli/Cli.java

+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/partition/IoTDBPartitionInheritStrategyIT.java
@@ -56,7 +56,7 @@ public class IoTDBPartitionInheritStrategyIT {
   private static final int testReplicationFactor = 1;
   private static final int testSeriesSlotNum = 1000;
   private static final long testTimePartitionInterval = 604800000;
-  private static final double testDataRegionPerDataNode = 5.0;
+  private static final int testDataRegionPerDataNode = 5;
 
   private static final String database = "root.database";
   private static final int seriesPartitionSlotBatchSize = 100;
